# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Airavate Backend API
  description: |
    Scalable production-grade backend API for Airavate project.
    
    ## Current Status
    This API is in initial setup phase. Currently only basic health check and info endpoints are available.
    
    ## Future Features
    - JWT Authentication (ready for Authentik integration)
    - User Management
    - Payment Processing (Razorpay integration)
    - Audit Logging
    - Rate Limiting
    - CORS Support
    
  version: 1.0.0
  contact:
    name: Airavate Team
    email: admin@airavate.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.airavate.com
    description: Production server (future)
paths:
...
  /api/v1/users/{id}:
    get:
      tags:
        - Users (Coming Soon)
      summary: Get user by ID endpoint (placeholder)
      description: Get user by ID endpoint - Coming soon
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
      responses:
        '501':
          description: Not implemented yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotImplementedResponse'

    put:
      tags:
        - Users (Coming Soon)
      summary: Update user endpoint (placeholder)
      description: Update user endpoint - Coming soon
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
      responses:
        '501':
          description: Not implemented yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotImplementedResponse'

    delete:
      tags:
        - Users (Coming Soon)
      summary: Delete user endpoint (placeholder)
      description: Delete user endpoint - Coming soon
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
      responses:
        '501':
          description: Not implemented yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotImplementedResponse'
components:
  schemas:
...
    NotImplementedResponse:
      type: object
      properties:
        message:
          type: string
          example: "This endpoint is coming soon"
        body:
          type: object
          nullable: true
        params:
          type: object
          nullable: true
        query:
          type: object
          nullable: true
...