# src/routes/user.routes.ts
import { Router, Request, Response } from 'express';

const router = Router();

// Placeholder user routes
router.get('/', (req: Request, res: Response) => {
  res.status(501).json({
    message: 'Get all users endpoint - Coming soon',
    query: req.query,
  });
});

router.get('/:id', (req: Request, res: Response) => {
  res.status(501).json({
    message: 'Get user by ID endpoint - Coming soon',
    params: req.params,
  });
});

router.post('/', (req: Request, res: Response) => {
  res.status(501).json({
    message: 'Create user endpoint - Coming soon',
    body: req.body,
  });
});

router.put('/:id', (req: Request, res: Response) => {
  res.status(501).json({
    message: 'Update user endpoint - Coming soon',
    params: req.params,
    body: req.body,
  });
});

router.delete('/:id', (req: Request, res: Response) => {
  res.status(501).json({
    message: 'Delete user endpoint - Coming soon',
    params: req.params,
  });
});

export default router;